+++ src/renderer/shader.cpp
@@ -5,7 +5,7 @@
 #include <sstream>
 #include <cstring>
 
-static void CheckShaderError(uint32_t shader, uint32_t flag, bool isProgram, String msg)
+static void CheckShaderError(int32_t shader, int32_t flag, bool isProgram, String msg)
 {
     int success = 0;
     char error[1024] = {0};
@@ -36,16 +36,16 @@ static void CheckShaderError(uint32_t shader, uint32_t flag, bool isProgram, Str
 
 } // checkShaderError
 
-static uint32_t CreateShader(const char *shaderSource, GLenum shaderType)
+static int32_t CreateShader(const char *shaderSource, GLenum shaderType)
 {
-    gl_call(uint32_t shader = glad_glCreateShader(shaderType));
+    gl_call(int32_t shader = glad_glCreateShader(shaderType));
     if (shader == 0)
     {
         std::cout << "Failed to create shader!\n";
         exit(0);
     }
 
-    int sourceLength = strlen(shaderSource);
+    int sourceLength = int(strlen(shaderSource));
     gl_call(glShaderSource(shader, 1, &shaderSource, &sourceLength));
 
     String error = "\nError: Shader compilation failed => ";
@@ -56,7 +56,7 @@ static uint32_t CreateShader(const char *shaderSource, GLenum shaderType)
     return shader;
 } // CreateShader
 
-static uint32_t genShader(const String &vertex_shader, const String &frag_shader)
+static int32_t genShader(const String &vertex_shader, const String &frag_shader)
 {
 
     auto readShader = [](const String &path) -> String {
@@ -67,11 +67,11 @@ static uint32_t genShader(const String &vertex_shader, const String &frag_shader
         return shader.str();
     };
 
-    uint32_t vtx = CreateShader(readShader(vertex_shader).c_str(), GL_VERTEX_SHADER);
-    uint32_t frg = CreateShader(readShader(frag_shader).c_str(), GL_FRAGMENT_SHADER);
+    int32_t vtx = CreateShader(readShader(vertex_shader).c_str(), GL_VERTEX_SHADER);
+    int32_t frg = CreateShader(readShader(frag_shader).c_str(), GL_FRAGMENT_SHADER);
 
     // Create program
-    gl_call(uint32_t program = glad_glCreateProgram());
+    gl_call(int32_t program = glad_glCreateProgram());
     gl_call(glad_glAttachShader(program, vtx));
     gl_call(glad_glAttachShader(program, frg));
 
@@ -131,74 +131,74 @@ void Shader::useProgram(const String &name)
 
 void Shader::setInteger(const String &name, int val)
 {
-    gl_call(uint32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
+    gl_call(int32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
     gl_call(glad_glUniform1i(loc, val));
 
 } // setUniform1i
 
 void Shader::setFloat(const String &name, float val)
 {
-    gl_call(uint32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
+    gl_call(int32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
     gl_call(glad_glUniform1f(loc, val));
 
 } // setUniform1i
 
 void Shader::setVec2f(const String &name, const float *v)
 {
-    gl_call(uint32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
+    gl_call(int32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
     gl_call(glad_glUniform2f(loc, v[0], v[1]));
 } // setUnifom3f
 
 void Shader::setVec3f(const String &name, const float *v)
 {
-    gl_call(uint32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
+    gl_call(int32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
     gl_call(glad_glUniform3f(loc, v[0], v[1], v[2]));
 } // setUnifom3f
 
 void Shader::setVec4f(const String &name, const float *v)
 {
-    gl_call(uint32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
+    gl_call(int32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
     gl_call(glad_glUniform4f(loc, v[0], v[1], v[2], v[3]));
 } // setUnifom3f
 
 void Shader::setMatrix3f(const String &name, const float *mat)
 {
-    gl_call(uint32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
+    gl_call(int32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
     gl_call(glad_glUniformMatrix3fv(loc, 1, GL_FALSE, mat));
 } //setUniformMatrix4f
 
 void Shader::setMatrix4f(const String &name, const float *mat)
 {
-    gl_call(uint32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
+    gl_call(int32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
     gl_call(glad_glUniformMatrix4fv(loc, 1, GL_FALSE, mat));
 } //setUniformMatrix4f
 
-void Shader::setIntArray(const String &name, const int *ptr, size_t N)
+void Shader::setIntArray(const String &name, const int *ptr, int32_t N)
 {
-    gl_call(uint32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
+    gl_call(int32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
     gl_call(glad_glUniform1iv(loc, N, ptr));
 }
 
-void Shader::setFloatArray(const String &name, const float *ptr, size_t N)
+void Shader::setFloatArray(const String &name, const float *ptr, int32_t N)
 {
-    gl_call(uint32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
+    gl_call(int32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
     gl_call(glad_glUniform1fv(loc, N, ptr));
 }
 
-void Shader::setVec2fArray(const String &name, const float *ptr, size_t N)
+void Shader::setVec2fArray(const String &name, const float *ptr, int32_t N)
 {
-    gl_call(uint32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
+    gl_call(int32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
     gl_call(glad_glUniform2fv(loc, N, ptr));
 }
 
-void Shader::setVec3fArray(const String &name, const float *ptr, size_t N)
+void Shader::setVec3fArray(const String &name, const float *ptr, int32_t N)
 {
-    gl_call(uint32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
+    gl_call(int32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
     gl_call(glad_glUniform3fv(loc, N, ptr));
 }
 
-void Shader::setMat3Array(const String &name, const float *ptr, size_t N)
+void Shader::setMat3Array(const String &name, const float *ptr, int32_t N)
 {
-    gl_call(uint32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
+    gl_call(int32_t loc = glad_glGetUniformLocation(program_used, name.c_str()));
     gl_call(glad_glUniformMatrix3fv(loc, N, true, ptr));
 }
